==> 2_clean_data.Rmd <==
---
title: "COPE Commonality Analysis"
author: "Arielle Smith and Laura Jans"
date: "2025-08-25"
output: html_document
---

```{r setup}
#Load required packages

library(groundhog)
groundhog_date <- "2025-04-10"
meta.groundhog(groundhog_date)

pkgs_cleaning <- c("tidyverse", "missForest", "forcats")
groundhog.library(pkgs_cleaning, groundhog_date)

#Read in data
cope_full_data <- readRDS("cleaned_cope_data_randomized.rds")
```

```{r data cleaning}
#Select necessary columns
cope_ca_data <- cope_full_data %>% 
  select(contains(c("b_dem", "b_bhs", "b_shs", "b_cdi", "b_gad", "b_drs", "b_cts", "b_pfs", "pi_bhs", "pi_shs", "f1_bhs", "f1_shs", "f1_cdi", "f1_gad", "f1_drs", "f1_cts", "condition"))) %>%
#Adding cdi sums (means already exist for all the other variables - which is what we need for our analyses!)
  mutate(b_cdi_sum = b_cdi_1 + b_cdi_2 + b_cdi_3 + b_cdi_4 + b_cdi_5 + b_cdi_6 + b_cdi_7 + b_cdi_8 + b_cdi_9 + b_cdi_10 + b_cdi_11 + b_cdi_12, 
         f1_cdi_sum = f1_cdi_1 + f1_cdi_2 + f1_cdi_3 + f1_cdi_4 + f1_cdi_5 + f1_cdi_6 + f1_cdi_7 + f1_cdi_8 + f1_cdi_9 + f1_cdi_10 + f1_cdi_11 + f1_cdi_12) %>% 
#Redoing baseline agency mean due to error in original data (all items were included instead of the subscale)
  mutate(b_shs_mean = (b_shs_1 + b_shs_3 + b_shs_5) / 3) %>% 
#Confusing that the program feedback scale was labelled with the abbreviation for baseline rather than post-intervention, so changing it
  rename_with(~ gsub("^b_pfs", "pi_pfs", .), starts_with("b_pfs")) %>% 
#Remove open-ended text variables (for demographics) because will not give missForest meaningful info
  select(-ends_with("_text"))

#Includes the demographics, predictor variables, outcome variables, and the condition. Condition is just labeled 0, 1, and 2, but looking at the COPE baseline survey flow and parent study code, have gathered that:
#0 = control (Sharing Feelings)
#1 = intervention (Project Personality)
#2 = intervention (ABC Project)

#Make a new variable to indicate SSI completion - same method as code from parent project:
cope_ca_data <- cope_ca_data %>% 
  mutate(pi_dropout = is.na(pi_pfs_1))
#pi_dropout stands for post-intervention dropout

#Select only the intervention participants who completed their SSI
cope_ca_ssi_completers <- cope_ca_data %>% 
  filter(condition == 1 | condition == 2,
         pi_dropout == 0)

#Relevel condition variable now that control participants have been removed
cope_ca_ssi_completers$condition <- factor(cope_ca_ssi_completers$condition,
                                          levels = c("1", "2"))

#Remove SSI completers without complete baseline data (we are not imputing these variables)
cope_ca_ssi_completers <- cope_ca_ssi_completers %>%
  drop_na(b_cdi_sum, b_gad_mean, b_drs_month, b_cts_rs_mean)
```

```{r data imputation}
#8/14/25: We originally planned to impute missing data using "missForest". We then ran initial analyses 
#using the resulting imputed data and found similar results as we did when we reran analyses on complete 
#cases. Given limitations of "missForest()", we ultimately decided not to use it (and used complete cases 
#instead). Thus, the following code related to "missForest" is commented out.

#Change numeric-coded and integer-coded categorical variables into factors and make cope_ca_ssi_completers 
#a data.frame for "missForest" to work
#cope_ca_ssi_completers <- as.data.frame(cope_ca_ssi_completers) %>% 
#  mutate(across(c(pi_dropout, b_drs_month, f1_drs_month, 
#                  starts_with(c("b_dem_gender", "b_dem_sex", "b_dem_rom", "b_dem_race"))), factor))

#Since "missForest" can only work with categorical variables with a maximum of 53 levels, need to lump 
#together factors. Decided to lump those with less than 10 responses.
#cope_ca_ssi_completers <- cope_ca_ssi_completers %>%
#  mutate(across(starts_with("b_dem"),
#                ~ fct_lump_min(., 10)))

#When running "missForest()" on 8/14/25 on macOS (see "2025.08.17_sessionInfo_for_cleaning_script.txt"),
#we got an "incorrect cutoff specified" error. Although we could not reproduce this error on Windows 11, 
#we resolved it on macOS using the code below, which removes factor variables with levels other than 2.
#factors_not_two_levels <- names(Filter(function(x) is.factor(x) && nlevels(x) != 2, cope_ca_ssi_completers))
#factors_not_two_levels

#cope_ca_ssi_completers_2 <- cope_ca_ssi_completers %>%
#  select(-all_of(factors_not_two_levels))

#set.seed(12345) #For reproducibility
#cope_ca_data_imputed <- missForest(cope_ca_ssi_completers_2)

#cope_ca_data_imputed$OOBerror
#NRMSE = 0.21. This is the error term for the continuous variables.
#PFC = 0.04. This is the error term for the categorical variables.

#Turn the imputed data into its own object
#cope_ca_data_imputed <- as.data.frame(cope_ca_data_imputed$ximp)

#Export as .csv
#write.csv(cope_ca_data_imputed, file='cope_ca_data_imputed.csv')
```

```{r create new data frame of complete cases}
selected_columns <- c("b_bhs", "b_shs", "b_cdi", "pi_pfs", "pi_bhs", "pi_shs", "f1_bhs", "f1_shs", "f1_cdi")

cope_ca_data_nonimputed <- cope_ca_ssi_completers %>%
  drop_na(contains(selected_columns))
#885 observations

#Export as .csv
write.csv(cope_ca_data_nonimputed, file='cope_ca_data_nonimputed.csv')
```

```{r pull session info}
sink(file = "2025.08.25_sessionInfo_for_cleaning_script.txt")
sessionInfo()
sink()
```
