==> 3_run_analysis.Rmd <==
---
title: "Commonality Analysis Code_2"
author: "Laura Jans and Arielle Smith"
date: "2025-08-21"
output: html_document
editor_options: 
  chunk_output_type: inline
---

# Loading Cleaned Data

```{r setup, echo=TRUE, message=FALSE, warning=FALSE}
#Load required packages

library(groundhog)
groundhog_date <- "2025-04-10"
meta.groundhog(groundhog_date)

pkgs_analysis <- c("tidyverse", "yhat", "boot", "car", "lmtest", "olsrr")
groundhog.library(pkgs_analysis, groundhog_date)

#Source custom functions
source("0_define_functions.R")

#Read in data from data cleaning and pre-processing code (to avoid re-running data prep every time)
cope_ca_data_nonimputed <- read.csv("cope_ca_data_nonimputed.csv")

set.seed(6804275)
```


# Creating Difference Scores

## Dependent Variable
```{r}
#Creating difference scores for depression (pre-SSI to 3M)
cope_ca_data_nonimputed <- cope_ca_data_nonimputed %>% 
  mutate(difference_score_cdi = f1_cdi_sum - b_cdi_sum)
```

## Independent Variables

```{r}
#Creating difference scores for hopelessness (pre-SSI to post-SSI) 
cope_ca_data_nonimputed <- cope_ca_data_nonimputed %>% 
  mutate(difference_score_bhs = pi_bhs_mean - b_bhs_mean)

#Creating difference scores for agency (pre-SSI to post-SSI) in both the imputed and non-imputed datasets 
cope_ca_data_nonimputed <- cope_ca_data_nonimputed %>% 
  mutate(difference_score_shs = pi_shs_mean - b_shs_mean)
```

# Running Regressions (Nonimputed)

```{r, results = "asis", include=T, echo = T, message=FALSE, warning=FALSE}

# Subset data based on condition

cope_ca_data_nonimputed_PP <- subset(cope_ca_data_nonimputed, condition == 1)
cope_ca_data_nonimputed_ABC <- subset(cope_ca_data_nonimputed, condition == 2)

# Run main regressions

# Project Personality, Nonnonimputed

cdi_PP_nonimputed <- lm(difference_score_cdi ~ difference_score_bhs + difference_score_shs + pi_pfs_mean, data = cope_ca_data_nonimputed_PP)
summary(cdi_PP_nonimputed)

# ABC Project, Nonnonimputed

cdi_ABC_nonimputed <- lm(difference_score_cdi ~ difference_score_bhs + difference_score_shs + pi_pfs_mean, data = cope_ca_data_nonimputed_ABC)
summary(cdi_ABC_nonimputed)
```
Note: We decided to use mean for the PFS instead of component scores.


# Commonality Analysis (Nonimputed)

```{r, results = "asis", include=T, echo = T, message=FALSE, warning=FALSE}

# First, let's look at the r^2 for each model

summary(cdi_PP_nonimputed)$r.squared
summary(cdi_ABC_nonimputed)$r.squared

# Decompose r^2 # Let's just try this once for the model predicting follow-up depression symptoms for Project Personality 

regOut_cdi_PP <- calc.yhat(cdi_PP_nonimputed, prec = 6)

# Bootstrap 1000 times to get CIs

set.seed(6804275)
boot.regOut_cdi_PP <-boot(cope_ca_data_nonimputed_PP, boot.yhat.prec, 1000, lmOut = cdi_PP_nonimputed, regrout0 = regOut_cdi_PP, prec = 6)
boot.regOut_cdi_PP_result <- booteval.yhat(regOut_cdi_PP, boot.regOut_cdi_PP, bty="perc", prec = 6)
print(boot.regOut_cdi_PP_result$combCIaps)
print(boot.regOut_cdi_PP_result$combCIpm["b"])

# Regress depression onto each predictor individually to get total effects
summary(lm(difference_score_cdi ~ difference_score_bhs, data = cope_ca_data_nonimputed_PP))
summary(lm(difference_score_cdi ~ difference_score_shs, data = cope_ca_data_nonimputed_PP))
summary(lm(difference_score_cdi ~ pi_pfs_mean, data = cope_ca_data_nonimputed_PP))

# Project ABC, Depression

regOut_cdi_ABC <- calc.yhat(cdi_ABC_nonimputed, prec = 6)

set.seed(6804275)
boot.regOut_cdi_ABC <- boot(cope_ca_data_nonimputed_ABC, boot.yhat.prec, 1000, lmOut = cdi_ABC_nonimputed, regrout0 = regOut_cdi_ABC, prec = 6)
boot.regOut_cdi_ABC_result <- booteval.yhat(regOut_cdi_ABC, boot.regOut_cdi_ABC, bty = "perc", prec = 6)
print(boot.regOut_cdi_ABC_result$combCIaps)
print(boot.regOut_cdi_ABC_result$combCIpm["b"])

summary(lm(difference_score_cdi ~ difference_score_bhs, data = cope_ca_data_nonimputed_ABC))
summary(lm(difference_score_cdi ~ difference_score_shs, data = cope_ca_data_nonimputed_ABC))
summary(lm(difference_score_cdi ~ pi_pfs_mean, data = cope_ca_data_nonimputed_ABC))
```

# Regression assumptions (Nonimputed)

```{r, results = "asis", include=T, echo = T, message=FALSE, warning=FALSE}

# Multicollinearity
sqrt(vif(cdi_PP_nonimputed)) # should be less than 2
sqrt(vif(cdi_ABC_nonimputed))
#met for all models

# Linearity of Data
plot(cdi_PP_nonimputed, 1)
plot(cdi_ABC_nonimputed, 1)
#met for all models (I think)

# Predictors are Independent
durbinWatsonTest(cdi_PP_nonimputed)
durbinWatsonTest(cdi_ABC_nonimputed)
#met for all models except for cdi_PP_nonimputed (p = 0.008)

# Homoscedasticity
##plot
plot(cdi_PP_nonimputed, 3)
plot(cdi_ABC_nonimputed, 3)
## original Breusch-Pagan test
ncvTest(cdi_PP_nonimputed)
ncvTest(cdi_ABC_nonimputed)
# Half met the assumption. These models did not: drs_PP_nonimputed, cdi_ABC_nonimputed, anxiety_ABC_nonimputed, cts_ABC_nonimputed
## studentized Breusch-Pagan test
bptest(cdi_PP_nonimputed)
bptest(cdi_ABC_nonimputed)
#same outcome

# Normality of Residual Errors
ols_plot_resid_qq(cdi_PP_nonimputed)
ols_plot_resid_qq(cdi_ABC_nonimputed)
# use Kolmogorov-Smirnov test
ols_test_normality(cdi_PP_nonimputed)
ols_test_normality(cdi_ABC_nonimputed)
# the only model that meets this assumption is anxiety_PP_nonimputed
```


```{r pull session info}
sink(file = "2025.08.21_sessionInfo_for_analysis_script.txt")
sessionInfo()
sink()
```
